SL-54: remove unused dependencies and functions
