--- src/data/inventoryData.js
@@ -1,25 +1,0 @@
-const groceries = [
-  {
-    id: 1,
-    item: 'Milk',
-    quantity: 2,
-    doe: '2023-11-06',
-    category: 'Dairy',
-  },
-  {
-    id: 2,
-    item: 'Eggs',
-    quantity: 1,
-    doe: '2023-11-1',
-    category: 'Dairy',
-  },
-  {
-    id: 3,
-    item: 'Bread',
-    quantity: 2,
-    doe: '2023-11-05',
-    category: 'Bakery',
-  },
-];
-
-export { groceries };
